Refactor R3: - removed deprecated robot-side GUI - move system_stat to a new folder - r3_robot is only responsible for reformatting & redirecting msgs to the server
